requirements.txt
peewee==2.8.0
psycopg2==2.6.1
wheel==0.26.0
python-dotenv==0.3.0
ordereddict==1.1
click==6.2
itsdangerous==0.24
Flask==0.10.1
Werkzeug==0.11.3
PyJWT==1.4.0
httplib2==0.9.2
socialauth==0.2.0
